Uploaded corrected files
In file group.cpp fixed line 145 and add 147 to support level less then 0.
In file playerMananger.cpp add to line 149 condition to support both levels are less then 0.
We still have a problem with the floating point. 
In test 2 line 7749, the hash-table returns wrong item and from there forward we fail in the test.